Better Implementation of Solving Groebner Basis
Wrote out the code which utilizes rationalpoints.jl. Ran a few tests and it should work fine, but note that rationalpoints.jl may return either multiple solutions or no solutions whatsoever (even if a potential solution is obvious).  The file rationapoints.jl is just for reference and could potentially be removed since a modified version of it is in NewtonsMethod.jl (necessary for it to run).